core_modules/UserService/requirements.txt
# UserService dependencies
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic[email]==2.3.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0             # used directly — passlib is unmaintained
python-dotenv==1.0.0
httpx==0.25.0          # async HTTP to call Middleware + IngestionEngine
twilio==8.5.0
sendgrid==6.10.0
python-multipart==0.0.6
